Sync with Prusa i3 branch
I modified the Prusa i3 files to suit my needs. I'm now using the same
inc folder structure. The configuration file is similar to Prusa i3,
with added bearings, belt information, and pulley information. I am
contemplating moving all this to a seperate file for easier portability.
The X-end is updated with my style of LM8UU holders, as well as
bushing.scad
rod_motor_holder.scad is now using the new style of bearing variables
for the idler. You can seperately define the idler bearings used for the
Y and X axis. (this is so you can have bearing mount screw that fit the
T-Slots).
I may do away with the Chamfer function for something more portable, but
it is an elegant solution as compared to using a bunch of math calls.
You can see the non-chamfer method in bushing.scad and it's not pretty.
(I probably messed up in my math too)